Refactor: extract git commands into own package
