+++ robots/cmd/test-label-analyzer/Makefile
@@ -5,7 +5,7 @@ format:
 	gofmt -w .
 
 test:
-	go test -v ./... ../../pkg/test-label-analyzer/...
+	go test -v ./... ../../pkg/test-label-analyzer/... ../../pkg/git/...
 
 build:
 	go build -v ./...
