+++ robots/cmd/test-label-analyzer/cmd/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     importpath = "kubevirt.io/project-infra/robots/cmd/test-label-analyzer/cmd",
     visibility = ["//visibility:public"],
     deps = [
+        "//robots/pkg/git:go_default_library",
         "//robots/pkg/test-label-analyzer:go_default_library",
         "//robots/pkg/test-report:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
@@ -27,6 +28,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
+        "//robots/pkg/git:go_default_library",
         "//robots/pkg/test-label-analyzer:go_default_library",
         "@com_github_onsi_ginkgo_v2//:go_default_library",
         "@com_github_onsi_gomega//:go_default_library",
