+++ robots/pkg/git/BUILD.bazel
@@ -2,24 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "config.go",
-        "ginkgo.go",
-        "stats.go",
-        "time.go",
-    ],
-    importpath = "kubevirt.io/project-infra/robots/pkg/test-label-analyzer",
+    srcs = ["git.go"],
+    importpath = "kubevirt.io/project-infra/robots/pkg/git",
     visibility = ["//visibility:public"],
-    deps = ["//robots/pkg/git:go_default_library"],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "stats_test.go",
-        "test_suite_test.go",
-        "time_test.go",
-    ],
+    srcs = ["git_test.go"],
     embed = [":go_default_library"],
     deps = [
         "@com_github_onsi_ginkgo_v2//:go_default_library",
